requirements.txt
langgraph
langchain-community
langchain
langchain-openai
google-generativeai
python-dotenv
Pillow
langchain_groq
PyMuPDF
aiofiles
langchain-ollama
langchain-tavily
bs4
streamlit
playwright



